Merge Develop to main

+++ src/modules/order-management/orderManagement.service.ts
@@ -25,13 +25,41 @@ export class OrderManagerService {
   ) {}
 
   public async createOrder(body: OrderMngDto): Promise<OrderManagementModel> {
-    await this.updateDeviceInOrder(body?.items);
+    await this.updateDeviceInOrder(body.items);
     const newOrder = new this.orderManagementModel({
       ...body,
       regDt: moment(new Date()).format('YYYY-MM-DD HH:mm:ss'),
     });
 
     return newOrder.save();
+  }
+
+  public async updateOrderById(
+    id: string,
+    body: OrderMngDto,
+  ): Promise<OrderManagementModel | boolean> {
+    const objectId = new Types.ObjectId(id);
+    const orderById = await this.getOrderById(id);
+
+    const updateDevice = await this.updateDeviceInOrder(
+      body?.items,
+      orderById?.items?.map((x) => ({
+        device: x.device.id as string,
+        details: x.details,
+      })) as ItemDto[],
+    );
+
+    if (updateDevice) {
+      const newOrder = await this.orderManagementModel.findOneAndUpdate(
+        {
+          _id: objectId,
+        },
+        { ...body, modDt: moment(new Date()).format('YYYY-MM-DD HH:mm:ss') },
+        { new: true },
+      );
+      return newOrder as OrderManagementModel;
+    }
+    return false;
   }
 
   public async getAllOrder(query: QueryOrderDto): Promise<ListOrderDto> {
@@ -131,47 +159,24 @@ export class OrderManagerService {
     }
   }
 
-  public async updateOrderById(
-    id: string,
-    body: OrderMngDto,
-  ): Promise<OrderManagementModel> {
-    const objectId = new Types.ObjectId(id);
-    const orderById = await this.getOrderById(id);
-
-    await this.updateDeviceInOrder(
-      body?.items,
-      orderById?.items?.map((x) => ({
-        device: x.device?.id as string,
-        quantity: x?.quantity as number,
-      })) as ItemDto[],
-    );
-    const newOrder = await this.orderManagementModel.findOneAndUpdate(
-      {
-        _id: objectId,
-      },
-      { ...body, modDt: moment(new Date()).format('YYYY-MM-DD HH:mm:ss') },
-      { new: true },
-    );
-
-    return newOrder as OrderManagementModel;
-  }
-
-  public async deleteOrderById(id: string): Promise<OrderManagementModel> {
+  public async deleteOrderById(id: string): Promise<boolean> {
     try {
       const objectId = new Types.ObjectId(id);
-      const deleteOrder = await this.orderManagementModel.findOneAndDelete({
-        _id: objectId,
-      });
+      const orderById = await this.getOrderById(id);
+      const oldItems = orderById?.items.map((x) => ({
+        device: x.device.id as string,
+        details: x.details,
+      })) as ItemDto[];
 
-      await this.updateDeviceInOrder(
-        [],
-        deleteOrder?.items?.map((x) => ({
-          device: x?.device?.toString(),
-          quantity: x?.quantity as number,
-        })) as ItemDto[],
-      );
+      const updateResult = await this.updateDeviceInOrder([], oldItems, true);
 
-      return deleteOrder as OrderManagementModel;
+      if (updateResult) {
+        await this.orderManagementModel.findOneAndDelete({
+          _id: objectId,
+        });
+        return true;
+      }
+      return false;
     } catch (error) {
       throw new HttpException(
         'An error occurred while delete the order',
@@ -183,139 +188,60 @@ export class OrderManagerService {
   public async updateDeviceInOrder(
     newItems?: ItemDto[],
     oldItems?: ItemDto[],
-  ): Promise<any> {
-    let updateSuccess;
+    setInventoryOnly: boolean = false,
+  ): Promise<boolean> {
+    const newItemsMap = new Map();
+    const oldItemsMap = new Map();
 
-    const updateDeviceStatus = async (
-      deviceId: string,
-      quantity: number,
-      isNewItem: boolean,
-    ) => {
-      const device = await this.deviceManagementModel.findById(deviceId);
-      if (!device) {
-        console.error(`Device with ID ${deviceId} not found.`);
-        return false;
-      }
+    (newItems || []).forEach((item) => {
+      newItemsMap.set(item.device, new Set(item.details));
+    });
 
-      const updatedStatus = device?.status?.map((statusItem) => {
-        if (statusItem?.status === 'inventory') {
-          return {
-            status: statusItem?.status,
-            quantity: isNewItem
-              ? statusItem.quantity - quantity
-              : statusItem.quantity + quantity,
-          };
-        } else if (statusItem?.status === 'sold') {
-          return {
-            status: statusItem?.status,
-            quantity: isNewItem
-              ? statusItem.quantity + quantity
-              : statusItem.quantity - quantity,
-          };
+    (oldItems || []).forEach((item) => {
+      oldItemsMap.set(item.device, new Set(item.details));
+    });
+
+    try {
+      if (!setInventoryOnly && newItems) {
+        for (const item of newItems) {
+          await this.deviceManagementModel.updateMany(
+            {
+              _id: item.device,
+              'detail.id_device': { $in: item.details },
+            },
+            {
+              $set: { 'detail.$[elem].status': 'sold' },
+            },
+            {
+              arrayFilters: [{ 'elem.id_device': { $in: item.details } }],
+            },
+          );
         }
-        return statusItem;
-      });
-
-      const updatedDeviceData = { status: updatedStatus };
-
-      const updateDevice = await this.deviceManagementModel.findOneAndUpdate(
-        { _id: device._id },
-        { $set: updatedDeviceData },
-        { new: true },
-      );
-
-      return !!updateDevice;
-    };
-
-    const checkOrderQuantity = async (
-      deviceId: string,
-      orderQuantity: number,
-    ) => {
-      const device = await this.deviceManagementModel.findById(deviceId);
-      const inventoryStatus = device?.status?.find(
-        (statusItem) => statusItem?.status === 'inventory',
-      );
-      if (orderQuantity > Number(inventoryStatus?.quantity)) {
-        throw new HttpException(
-          `Order quantity (${orderQuantity}) exceeds inventory quantity (${inventoryStatus?.quantity}) of the device.`,
-          HttpStatus.BAD_REQUEST,
-        );
       }
-      return true;
-    };
 
-    if (oldItems) {
-      // update order and device
-      const allDeviceIds = new Set([
-        ...oldItems.map((item) => item.device),
-        ...(newItems?.map((item) => item.device) || []),
-      ]);
-
-      for (const deviceId of allDeviceIds) {
-        const oldItem = oldItems.find((item) => item.device === deviceId);
-        const newItem = newItems?.find((item) => item.device === deviceId);
-
-        if (newItem?.device === oldItem?.device) {
-          if (newItem?.quantity !== oldItem?.quantity) {
-            const quantityDifference =
-              Number(newItem?.quantity) - Number(oldItem?.quantity);
-            const checkQuantityWhenOrder = await checkOrderQuantity(
-              deviceId,
-              Number(newItem?.quantity),
-            );
-            if (checkQuantityWhenOrder) {
-              const success = await updateDeviceStatus(
-                deviceId,
-                Math.abs(quantityDifference),
-                quantityDifference > 0,
+      if (oldItems && oldItems.length > 0) {
+        for (const oldItem of oldItems) {
+          const newDetailsSet = newItemsMap.get(oldItem.device) || new Set();
+          for (const detail of oldItem.details || []) {
+            if (setInventoryOnly || !newDetailsSet.has(detail)) {
+              await this.deviceManagementModel.updateOne(
+                {
+                  _id: oldItem.device,
+                  'detail.id_device': detail,
+                },
+                {
+                  $set: { 'detail.$.status': 'inventory' },
+                },
               );
-              if (success) updateSuccess = true;
             }
-          } else {
-            updateSuccess = true;
-          }
-        } else if (!newItem) {
-          const success = await updateDeviceStatus(
-            deviceId,
-            Number(oldItem?.quantity),
-            false,
-          );
-          if (success) updateSuccess = true;
-        } else if (!oldItem) {
-          const checkQuantityWhenOrder = await checkOrderQuantity(
-            deviceId,
-            Number(newItem?.quantity),
-          );
-          if (checkQuantityWhenOrder) {
-            const success = await updateDeviceStatus(
-              deviceId,
-              Number(newItem?.quantity),
-              true,
-            );
-            if (success) updateSuccess = true;
           }
         }
       }
-    } else {
-      // create order and update device
-      for (const item of newItems as ItemDto[]) {
-        if (item?.quantity > 0) {
-          const checkQuantityWhenOrder = await checkOrderQuantity(
-            item?.device,
-            Number(item?.quantity),
-          );
-          if (checkQuantityWhenOrder) {
-            const success = await updateDeviceStatus(
-              item?.device,
-              item?.quantity,
-              true,
-            );
-            if (success) updateSuccess = true;
-          }
-        }
-      }
-    }
 
-    return updateSuccess;
+      return true;
+    } catch (error) {
+      console.error('Error updating devices in order:', error);
+      return false;
+    }
   }
 }
